:zap: improve api call efficiency
