+++ .svelte-kit/ambient.d.ts
@@ -26,7 +26,6 @@
  * ```
  */
 declare module '$env/static/private' {
-	export const OPENAI_API_KEY: string;
 	export const TERM_PROGRAM: string;
 	export const NODE: string;
 	export const INIT_CWD: string;
@@ -82,7 +81,7 @@ declare module '$env/static/private' {
  * ```
  */
 declare module '$env/static/public' {
-	
+	export const PUBLIC_OPENAI_API_KEY: string;
 }
 
 /**
@@ -99,7 +98,6 @@ declare module '$env/static/public' {
  */
 declare module '$env/dynamic/private' {
 	export const env: {
-		OPENAI_API_KEY: string;
 		TERM_PROGRAM: string;
 		NODE: string;
 		INIT_CWD: string;
@@ -160,6 +158,7 @@ declare module '$env/dynamic/private' {
  */
 declare module '$env/dynamic/public' {
 	export const env: {
+		PUBLIC_OPENAI_API_KEY: string;
 		[key: `PUBLIC_${string}`]: string | undefined;
 	}
 }
